+++ appiumset/pom.xml
@@ -18,7 +18,7 @@
 <dependency>
     <groupId>io.appium</groupId>
     <artifactId>java-client</artifactId>
-    <version>7.0.0</version>
+    <version>7.3.0</version>
 </dependency>
 
 <dependency>
