+++ appiumset/src/test/java/Hanwhatribes_automation/Commonmethod.java
@@ -1,3 +1,4 @@
+package Hanwhatribes_automation;
 /*
  * package appiumset;
  *
