+++ appiumset/test-output/Suite/Device2.xml
@@ -1,30 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-
-<testsuite ignored="0" hostname="DESKTOP-H9VDVB7" failures="0" tests="1" name="Device2" time="2.441" errors="0" timestamp="2022-11-01T14:43:23 KST">
-  <testcase classname="appiumset._1_Deviceinfo" name="@BeforeMethod Deviceinfo" time="2.4">
+<testsuite ignored="0" hostname="DESKTOP-H9VDVB7" failures="0" tests="1" name="Device2" time="2.467" errors="0" timestamp="2022-11-03T14:37:32 KST">
+  <testcase classname="appiumset._1_Deviceinfo" name="@BeforeMethod Deviceinfo" time="2.425">
     <failure type="org.openqa.selenium.WebDriverException" message="Connection refused: no further information
 Build info: version: &amp;apos;3.141.59&amp;apos;, revision: &amp;apos;e82be7d358&amp;apos;, time: &amp;apos;2018-11-14T08:17:03&amp;apos;
 System info: host: &amp;apos;DESKTOP-H9VDVB7&amp;apos;, ip: &amp;apos;192.168.0.83&amp;apos;, os.name: &amp;apos;Windows 10&amp;apos;, os.arch: &amp;apos;amd64&amp;apos;, os.version: &amp;apos;10.0&amp;apos;, java.version: &amp;apos;17.0.4.1&amp;apos;
 Driver info: driver.version: AppiumDriver">
       <![CDATA[org.openqa.selenium.WebDriverException: Connection refused: no further information
 Build info: version: '3.141.59', revision: 'e82be7d358', time: '2018-11-14T08:17:03'
-
+System info: host: 'DESKTOP-H9VDVB7', ip: '192.168.0.83', os.name: 'Windows 10', os.arch: 'amd64', os.version: '10.0', java.version: '17.0.4.1'
 Driver info: driver.version: AppiumDriver
 at io.appium.java_client.remote.AppiumCommandExecutor.lambda$5(AppiumCommandExecutor.java:251)
 at java.base/java.util.Optional.orElseGet(Optional.java:364)
 at io.appium.java_client.remote.AppiumCommandExecutor.execute(AppiumCommandExecutor.java:250)
 at org.openqa.selenium.remote.RemoteWebDriver.execute(RemoteWebDriver.java:552)
-at io.appium.java_client.DefaultGenericMobileDriver.execute(DefaultGenericMobileDriver.java:42)
+at io.appium.java_client.DefaultGenericMobileDriver.execute(DefaultGenericMobileDriver.java:41)
 at io.appium.java_client.AppiumDriver.execute(AppiumDriver.java:1)
 at org.openqa.selenium.remote.RemoteWebDriver.startSession(RemoteWebDriver.java:213)
+at io.appium.java_client.AppiumDriver.startSession(AppiumDriver.java:336)
 at org.openqa.selenium.remote.RemoteWebDriver.<init>(RemoteWebDriver.java:131)
-at io.appium.java_client.DefaultGenericMobileDriver.<init>(DefaultGenericMobileDriver.java:38)
-at io.appium.java_client.AppiumDriver.<init>(AppiumDriver.java:84)
-at io.appium.java_client.AppiumDriver.<init>(AppiumDriver.java:94)
-
+at io.appium.java_client.DefaultGenericMobileDriver.<init>(DefaultGenericMobileDriver.java:37)
+at io.appium.java_client.AppiumDriver.<init>(AppiumDriver.java:88)
+at io.appium.java_client.AppiumDriver.<init>(AppiumDriver.java:98)
 at appiumset._1_Deviceinfo.Deviceinfo(_1_Deviceinfo.java:56)
-
 at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
 at java.base/java.util.concurrent.FutureTask.run(FutureTask.java:264)
 at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1136)
@@ -61,7 +59,7 @@ at org.openqa.selenium.remote.ProtocolHandshake.createSession(ProtocolHandshake.
 at io.appium.java_client.remote.AppiumCommandExecutor$1.createSession(AppiumCommandExecutor.java:186)
 at io.appium.java_client.remote.AppiumCommandExecutor.createSession(AppiumCommandExecutor.java:217)
 at io.appium.java_client.remote.AppiumCommandExecutor.execute(AppiumCommandExecutor.java:239)
-... 36 more
+... 37 more
 ... Removed 26 stack frames]]>
     </failure>
   </testcase> <!-- @BeforeMethod Deviceinfo -->
