+++ appiumset/test-output/old/Suite/methods.html
@@ -1,14 +1,12 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-
-<tr bgcolor="ea606c">  <td>22/11/01 14:43:23</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_0_Automation_Main.main()[pri:0, instance:appiumset._0_Automation_Main@485966cc]">main</td> 
-  <td>TestNG-tests-1@1219080126</td>   <td></td> </tr>
-<tr bgcolor="98ac89">  <td>22/11/01 14:43:20</td>   <td>-2404</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;_1_Deviceinfo.Deviceinfo(java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:appiumset._0_Automation_Main@485966cc]">&gt;&gt;Deviceinfo</td> 
-<td>&nbsp;</td>  <td>TestNG-tests-1@1219080126</td>   <td></td> </tr>
-<tr bgcolor="ea606c">  <td>22/11/01 14:43:23</td>   <td>1</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_0_Automation_Main.main()[pri:0, instance:appiumset._0_Automation_Main@20140db9]">main</td> 
-  <td>TestNG-tests-2@1027330158</td>   <td></td> </tr>
-<tr bgcolor="98ac89">  <td>22/11/01 14:43:20</td>   <td>-2404</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;_1_Deviceinfo.Deviceinfo(java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:appiumset._0_Automation_Main@20140db9]">&gt;&gt;Deviceinfo</td> 
-<td>&nbsp;</td>  <td>TestNG-tests-2@1027330158</td>   <td></td> </tr>
-
+<tr bgcolor="909dad">  <td>22/11/03 14:37:38</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_0_Automation_Main.main()[pri:0, instance:appiumset._0_Automation_Main@43dac38f]">main</td> 
+  <td>TestNG-tests-1@2079883433</td>   <td></td> </tr>
+<tr bgcolor="8edcc5">  <td>22/11/03 14:37:30</td>   <td>-7892</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;_1_Deviceinfo.Deviceinfo(java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:appiumset._0_Automation_Main@43dac38f]">&gt;&gt;Deviceinfo</td> 
+<td>&nbsp;</td>  <td>TestNG-tests-1@2079883433</td>   <td></td> </tr>
+<tr bgcolor="8edcc5">  <td>22/11/03 14:37:30</td>   <td>-7892</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;_1_Deviceinfo.Deviceinfo(java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String,java.lang.String)[pri:0, instance:appiumset._0_Automation_Main@6574a52c]">&gt;&gt;Deviceinfo</td> 
+<td>&nbsp;</td>  <td>TestNG-tests-2@1696837705</td>   <td></td> </tr>
+<tr bgcolor="909dad">  <td>22/11/03 14:37:32</td>   <td>-5457</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_0_Automation_Main.main()[pri:0, instance:appiumset._0_Automation_Main@6574a52c]">main</td> 
+  <td>TestNG-tests-2@1696837705</td>   <td></td> </tr>
 </table>
